+++ server/src/main/java/models/AuthToken.java
@@ -3,7 +3,7 @@ package models;
 import java.time.OffsetDateTime;
 import java.util.UUID;
 
-public record AuthToken(Long id, String username, UUID authToken, OffsetDateTime createdAt) implements Model {
+public record AuthToken(Integer id, String username, UUID authToken, OffsetDateTime createdAt) implements Model {
     public String getTableName(){
         return "AuthTokens";
     }
