+++ server/src/main/java/models/Game.java
@@ -1,14 +1,12 @@
 package models;
 
 public record Game(
-    Long id,
-    String gameName,
-    String whiteUsername,
-    String blackUsername,
-    String game
-) implements Model {
-        public String getTableName(){
+        Integer id,
+        String gameName,
+        String whiteUsername,
+        String blackUsername,
+        String game) implements Model {
+    public String getTableName() {
         return "Games";
     }
 }
-
